Add settings command
